testing changes
testing with a new writer as imagemagick is funky

changed the running of the run_pipeline to have 2 inputs, the first for the supernova name and the 2nd to be the name of the template to be used. omit the 'y' argument and just go with it.

the idea is to port the 3d generation to use pillow instead of imagemagick as the movie writer. have not got it to work yet.